+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ## 0.0.1
 
-* First Release of implementation of Android Sdk of Checkout Sheet Shopify 
+* First Release of implementation of Android Sdk of Checkout Sheet Shopify iOS Coming Soon
+## 0.0.2
+
+* iOS Support added using Shopify iOS checkout sdk version 3.1.2